+++ src/main/kotlin/online/danielstefani/paddy/stats/dto/TotalPower.kt
@@ -1,3 +1,3 @@
 package online.danielstefani.paddy.stats.dto
 
-data class TotalPower(val kWh: Float)
+data class TotalPower(val kWh: Double)
